[WFLY-13972] Separate testing of deprecated @SafeHtml into a separate package
